Mac build fixes.
This gets us pretty close. There are a few problems with libraries we depend
on (system/core/liblog and external/icu4c) that I have workarounds for, and
a problem with gtest that I haven't yet worked around that prevents us from
linking any of the tests. But this at least gives us a Mac dex2oat binary.

+++ src/runtime_support.cc
@@ -1215,11 +1215,11 @@ extern "C" uintptr_t artTraceMethodExitFromCode() {
   return lr;
 }
 
-uintptr_t artTraceMethodUnwindFromCode(Thread* self) {
+uint32_t artTraceMethodUnwindFromCode(Thread* self) {
   Trace* tracer = Runtime::Current()->GetTracer();
   TraceStackFrame trace_frame = self->PopTraceStackFrame();
   Method* method = trace_frame.method_;
-  uintptr_t lr = trace_frame.return_pc_;
+  uint32_t lr = trace_frame.return_pc_;
 
   tracer->LogMethodTraceEvent(self, method, Trace::kMethodTraceUnwind);
